+++ experiments/test_methodology_quick.json
@@ -1,45 +1,18 @@
 {
   "name": "Test new methodology",
   "kernels": [
-    "pnpoly",
-    "convolution"
+    "pnpoly"
   ],
   "GPUs": [
     "RTX_2070_SUPER"
   ],
   "resolution": 1e3,
   "curve_segment_factor": 0.05,
-  "cutoff_quantile": 0.95,
+  "cutoff_quantile": 0.98,
   "relative_to_baseline": true,
   "strategy_defaults": {
-    "repeats": 3,
-    "nums_of_evaluations": [
-      20,
-      40,
-      60,
-      80,
-      100,
-      120,
-      140,
-      160,
-      180,
-      200,
-      220,
-      240,
-      260,
-      280,
-      300,
-      320,
-      340,
-      360,
-      380,
-      400,
-      420,
-      440,
-      460,
-      480,
-      500
-    ],
+    "repeats": 100,
+    "minimum_number_of_evaluations": 20,
     "record_data": [
       "time",
       "GFLOP/s"
@@ -47,9 +20,25 @@
   },
   "strategies": [
     {
+      "name": "random_sample",
+      "strategy": "random_sample",
+      "display_name": "Random sampling",
+      "repeats": 500
+    },
+    {
       "name": "mls",
       "strategy": "mls",
       "display_name": "MLS"
+    },
+    {
+      "name": "greedy_mls",
+      "strategy": "greedy_mls",
+      "display_name": "Greedy MLS"
+    },
+    {
+      "name": "greedy_ils",
+      "strategy": "greedy_ils",
+      "display_name": "Greedy ILS"
     }
   ]
 }
