+++ experiments/test_methodology_quicker.json
@@ -1,0 +1,29 @@
+{
+  "name": "Test new methodology",
+  "kernels": [
+    "pnpoly"
+  ],
+  "GPUs": [
+    "RTX_2070_SUPER"
+  ],
+  "resolution": 1e3,
+  "curve_segment_factor": 0.05,
+  "cutoff_quantile": 0.98,
+  "relative_to_baseline": true,
+  "strategy_defaults": {
+    "repeats": 100,
+    "minimum_number_of_evaluations": 20,
+    "record_data": [
+      "time",
+      "GFLOP/s"
+    ]
+  },
+  "strategies": [
+    {
+        "name": "random_sample",
+        "strategy": "random_sample",
+        "display_name": "Random sampling",
+        "repeats": 200
+    }
+  ]
+}
